+++ src/phpcs-report/__tests__/worker-integration.test.ts
@@ -3,7 +3,6 @@ import * as child_process from 'child_process';
 import { Request } from '../request';
 import { ReportType } from '../response';
 import { WorkerPool } from '../worker-pool';
-import { StandardType } from '../../services/configuration';
 import { CancellationError } from 'vscode';
 import { MockCancellationToken } from '../../__mocks__/vscode';
 
@@ -53,7 +52,7 @@ describe('Worker/WorkerPool Integration', () => {
 				options: {
 					workingDirectory: __dirname,
 					executable: phpcsPath,
-					standard: StandardType.PSR12,
+					standard: 'PSR12',
 				},
 				data: null,
 			};
@@ -81,7 +80,7 @@ describe('Worker/WorkerPool Integration', () => {
 				options: {
 					workingDirectory: __dirname,
 					executable: phpcsPath,
-					standard: StandardType.PSR12,
+					standard: 'PSR12',
 				},
 				data: null,
 			};