+++ src/services/__tests__/diagnostic-updater.spec.ts
@@ -9,11 +9,7 @@ import {
 } from 'vscode';
 import { resolve as resolvePath } from 'path';
 import { CodeAction, CodeActionCollection } from '../../types';
-import {
-	Configuration,
-	LintAction,
-	StandardType,
-} from '../../services/configuration';
+import { Configuration, LintAction } from '../../services/configuration';
 import { WorkerPool } from '../../phpcs-report/worker-pool';
 import { DiagnosticUpdater } from '../diagnostic-updater';
 import {
@@ -96,7 +92,7 @@ describe('DiagnosticUpdater', () => {
 			executable: 'phpcs-test',
 			exclude: [],
 			lintAction: LintAction.Change,
-			standard: StandardType.PSR12,
+			standard: 'PSR12',
 		});
 		jest.mocked(mockWorker).execute.mockImplementation((request) => {
 			expect(request).toMatchObject({
@@ -104,7 +100,7 @@ describe('DiagnosticUpdater', () => {
 				options: {
 					workingDirectory: 'test-dir',
 					executable: 'phpcs-test',
-					standard: StandardType.PSR12,
+					standard: 'PSR12',
 				},
 			});
 
@@ -153,7 +149,7 @@ describe('DiagnosticUpdater', () => {
 			executable: 'phpcs-test',
 			exclude: [],
 			lintAction: LintAction.Change,
-			standard: StandardType.PSR12,
+			standard: 'PSR12',
 		});
 		jest.mocked(mockWorker).execute.mockImplementation((request) => {
 			expect(request).toMatchObject({
@@ -161,7 +157,7 @@ describe('DiagnosticUpdater', () => {
 				options: {
 					workingDirectory: 'test-dir',
 					executable: 'phpcs-test',
-					standard: StandardType.PSR12,
+					standard: 'PSR12',
 				},
 			});
 
@@ -186,7 +182,7 @@ describe('DiagnosticUpdater', () => {
 			executable: 'phpcs-test',
 			exclude: [new RegExp('.*/file/.*')],
 			lintAction: LintAction.Change,
-			standard: StandardType.PSR12,
+			standard: 'PSR12',
 		});
 
 		return diagnosticUpdater.update(document, LintAction.Force);
@@ -201,7 +197,7 @@ describe('DiagnosticUpdater', () => {
 			executable: 'phpcs-test',
 			exclude: [],
 			lintAction: LintAction.Save,
-			standard: StandardType.PSR12,
+			standard: 'PSR12',
 		});
 
 		return diagnosticUpdater.update(document, LintAction.Change);