+++ src/services/configuration.ts
@@ -11,16 +11,27 @@ import {
 import { UriMap } from '../common/uri-map';
 
 /**
- * An enum describing the values in the `phpcsCodeSniffer.standard` configuration.
+ * An enum describing the special parsing values in the `phpCodeSniffer.standard` configuration.
  */
-export enum StandardType {
+export enum SpecialStandardOptions {
+	/**
+	 * Disable linting for the document entirely.
+	 */
 	Disabled = 'Disabled',
+
+	/**
+	 * Skip passing a `--standard` to PHPCS and let it decide what standard to use.
+	 */
 	Default = 'Default',
-	PEAR = 'PEAR',
-	MySource = 'MySource',
-	Squiz = 'Squiz',
-	PSR1 = 'PSR1',
-	PSR12 = 'PSR12',
+
+	/**
+	 * Search for a custom standard file in the document's folder and parent folders.
+	 */
+	Automatic = 'Automatic',
+
+	/**
+	 * Use the contents of the `phpCodeSniffer.standardCustom` configuration.
+	 */
 	Custom = 'Custom',
 }
 
@@ -59,11 +70,11 @@ interface ParamsFromConfiguration {
 	executable: string;
 	exclude: RegExp[];
 	lintAction: LintAction;
-	standard: string;
+	standard: string | null;
 }
 
 /**
- * An interface descirinb the shape of a document's configuration.
+ * An interface describing the shape of a document's configuration.
  */
 export interface DocumentConfiguration {
 	/**
@@ -89,8 +100,23 @@ export interface DocumentConfiguration {
 	/**
 	 * The standard we should use when executing reports.
 	 */
-	standard: StandardType | string;
+	standard: string | null;
 }
+
+/**
+ * The type for the callback used when traversing workspace folders.
+ */
+type FolderTraversalCallback<T> = (folderUri: Uri) => Promise<T | false>;
+
+/**
+ * The valid filenames we look for when automatically searching for coding standards.
+ */
+export const AutomaticCodingStandardFilenames = [
+	'phpcs.xml',
+	'.phpcs.xml',
+	'phpcs.dist.xml',
+	'.phpcs.dist.xml',
+];
 
 /**
  * A class for reading our configuration.
@@ -138,7 +164,10 @@ export class Configuration {
 		}
 
 		// Read the configuration and filesystem to build our document's configuration.
-		const fromConfig = this.readConfiguration(document);
+		const fromConfig = await this.readConfiguration(
+			document,
+			cancellationToken
+		);
 		const fromFilesystem = await this.readFilesystem(
 			document,
 			fromConfig.autoExecutable,
@@ -173,49 +202,15 @@ export class Configuration {
 	}
 
 	/**
-	 * Reads the configuration for a document from the filesystem and resolves it.
-	 *
-	 * @param {TextDocument} document The document to read.
-	 * @param {boolean} findExecutable Indicates whether or not we should perform an executable search.
-	 * @param {CancellationToken} [cancellationToken] The optional token for cancelling the request.
-	 */
-	private async readFilesystem(
-		document: TextDocument,
-		findExecutable: boolean,
-		cancellationToken?: CancellationToken
-	): Promise<ParamsFromFilesystem> {
-		// The workspace folder for the document is our default working directory.
-		const workspaceFolder = this.getWorkspaceFolder(document);
-
-		// Prepare the parameters that come from the filesystem.
-		const fsParams: ParamsFromFilesystem = {
-			workingDirectory: workspaceFolder.fsPath,
-		};
-
-		// When an executable is requested we should attempt to populate the params with one.
-		if (findExecutable) {
-			const executable = findExecutable
-				? await this.findExecutable(
-						document.uri,
-						workspaceFolder,
-						cancellationToken
-				  )
-				: null;
-			if (executable) {
-				fsParams.workingDirectory = executable.workingDirectory;
-				fsParams.executable = executable.executable;
-			}
-		}
-
-		return fsParams;
-	}
-
-	/**
 	 * Reads the configuration for a document and returns the relevant data.
 	 *
 	 * @param {TextDocument} document The document to read.
+	 * @param {CancellationToken} [cancellationToken] The optional token for cancelling the request.
 	 */
-	private readConfiguration(document: TextDocument): ParamsFromConfiguration {
+	private async readConfiguration(
+		document: TextDocument,
+		cancellationToken?: CancellationToken
+	): Promise<ParamsFromConfiguration> {
 		const config = this.workspace.getConfiguration(
 			'phpCodeSniffer',
 			document
@@ -297,13 +292,22 @@ export class Configuration {
 			);
 		}
 
-		let standard = config.get<string>('standard');
-		if (standard === StandardType.Custom) {
-			standard = config.get<string>('standardCustom');
+		// We're going to parse the standard so that outside of this method
+		// we have a standard that can be easily passed to the worker.
+		const rawStandard = config.get<SpecialStandardOptions | string>(
+			'standard'
+		);
+		if (rawStandard === undefined) {
+			throw new Error(
+				'The extension has an invalid `phpCodeSniffer.standard` configuration.'
+			);
 		}
-		if (!standard) {
-			standard = StandardType.Disabled;
-		}
+		const standard = await this.parseStandard(
+			document,
+			rawStandard,
+			config.get<string>('standardCustom'),
+			cancellationToken
+		);
 
 		return {
 			autoExecutable,
@@ -312,6 +316,103 @@ export class Configuration {
 			lintAction,
 			standard,
 		};
+	}
+
+	/**
+	 * Parses the coding standard configuration options into a single string that
+	 * can be readily given to the worker without any other parsing.
+	 *
+	 * @param {TextDocument} document The document to read.
+	 * @param {SpecialStandardOptions|string} standard The special standard option or string literal to use.
+	 * @param {string} [customStandard] The string to use with the `Custom` special standard option.
+	 * @param {CancellationToken} [cancellationToken] The optional token for cancelling the request.
+	 */
+	private async parseStandard(
+		document: TextDocument,
+		standard: SpecialStandardOptions | string,
+		customStandard?: string,
+		cancellationToken?: CancellationToken
+	): Promise<string | null> {
+		// There are some special standard options that require some parsing.
+		switch (standard) {
+			// Linting will not be performed when the standard is null.
+			case SpecialStandardOptions.Disabled:
+				return null;
+
+			// No standard is passed when it is an empty string.
+			case SpecialStandardOptions.Default:
+				return '';
+
+			case SpecialStandardOptions.Custom:
+				if (!customStandard) {
+					throw new Error(
+						'The extension has an empty `phpCodeSniffer.standardCustom` configuration.'
+					);
+				}
+
+				return customStandard;
+
+			// Use the automatic standard discovery below when desired.
+			case SpecialStandardOptions.Automatic:
+				break;
+
+			// Any other standard options are just string literals to pass to the linter.
+			default:
+				return standard;
+		}
+
+		// We are only going to traverse as high as the workspace folder.
+		const workspaceFolder = this.getWorkspaceFolder(document);
+
+		const parsed = await this.traverseWorkspaceFolders(
+			document.uri,
+			workspaceFolder,
+			(uri) => this.findCodingStandardFile(uri),
+			cancellationToken
+		);
+		if (parsed === false) {
+			return null;
+		}
+
+		return parsed;
+	}
+
+	/**
+	 * Reads the configuration for a document from the filesystem and resolves it.
+	 *
+	 * @param {TextDocument} document The document to read.
+	 * @param {boolean} findExecutable Indicates whether or not we should perform an executable search.
+	 * @param {CancellationToken} [cancellationToken] The optional token for cancelling the request.
+	 */
+	private async readFilesystem(
+		document: TextDocument,
+		findExecutable: boolean,
+		cancellationToken?: CancellationToken
+	): Promise<ParamsFromFilesystem> {
+		// The workspace folder for the document is our default working directory.
+		const workspaceFolder = this.getWorkspaceFolder(document);
+
+		// Prepare the parameters that come from the filesystem.
+		const fsParams: ParamsFromFilesystem = {
+			workingDirectory: workspaceFolder.fsPath,
+		};
+
+		// When an executable is requested we should attempt to populate the params with one.
+		if (findExecutable) {
+			const executable = await this.traverseWorkspaceFolders(
+				document.uri,
+				workspaceFolder,
+				(uri) => this.findExecutableInFolder(uri),
+				cancellationToken
+			);
+
+			if (executable !== false) {
+				fsParams.workingDirectory = executable.workingDirectory;
+				fsParams.executable = executable.executable;
+			}
+		}
+
+		return fsParams;
 	}
 
 	/**
@@ -334,77 +435,108 @@ export class Configuration {
 			return this.workspace.workspaceFolders[0].uri;
 		}
 
-		// When we can't infer a path just use the directory of the document.
+		// When we can't infer a path just use the folder of the document.
 		return Uri.joinPath(document.uri, '..');
 	}
 
 	/**
-	 * Attempts to find an executable by traversing from the document's directory to the workspace folder.
+	 * Traverses from the document's folder to the workspace folder, executing
+	 * a callback on each Uri until the caller finds what they are looking for.
 	 *
-	 * @param {Uri} documentUri The URI of the document to find an executable for.
-	 * @param {Uri} workspaceFolder The URI of the workspace folder for the document.
+	 * @param {Uri} documentUri The Uri of the document we are traversing from.
+	 * @param {Uri} workspaceFolder The workspace folder that is the highest we should traverse.
+	 * @param {FolderTraversalCallback} callback The callback to execute on each Uri in the traversal.
 	 * @param {CancellationToken} [cancellationToken] The optional token for cancelling the request.
 	 */
-	private async findExecutable(
+	private async traverseWorkspaceFolders<T>(
 		documentUri: Uri,
 		workspaceFolder: Uri,
+		callback: FolderTraversalCallback<T>,
 		cancellationToken?: CancellationToken
-	): Promise<ExecutablePath | null> {
+	): Promise<T | false> {
 		// Where we start the traversal will depend on the scheme of the document.
-		let directory: Uri;
+		let folder: Uri;
 		switch (documentUri.scheme) {
 			// Untitled files have no path and should just check the workspace folder.
 			case 'untitled':
-				directory = workspaceFolder;
+				folder = workspaceFolder;
 				break;
 
-			// Real files will traverse from their directory to the workspace folder.
+			// Real files will traverse from their folder to the workspace folder.
 			case 'file':
-				directory = Uri.joinPath(documentUri, '..');
+				folder = Uri.joinPath(documentUri, '..');
 				break;
 
 			// Since we can't execute the binary in any other scheme there's nothing to do.
 			default:
-				return null;
+				return false;
 		}
 
-		// We're going to traverse from the file's directory to the workspace
-		// folder looking for an executable that can be used in the worker.
-		while (directory.path !== '/') {
+		// Only traverse as far as the workspace folder. We don't
+		// want to accidentally check folders outside of it.
+		while (folder.path !== '/') {
 			// When the request is cancelled, we don't want to keep looking.
 			if (cancellationToken?.isCancellationRequested) {
 				throw new CancellationError();
 			}
 
-			const found = await this.findExecutableInDirectory(directory);
-			if (found) {
+			// Let the caller decide whether or not the given
+			// Uri is what they're looking for and return
+			// whatever the caller wants to consume.
+			const found = await callback(folder);
+			if (found !== false) {
 				return found;
 			}
 
 			// Stop once we reach the workspace folder.
-			if (directory.toString() === workspaceFolder.toString()) {
+			if (folder.toString() === workspaceFolder.toString()) {
 				break;
 			}
 
-			// Move to the parent directory and check again.
-			directory = Uri.joinPath(directory, '..');
+			// Move to the parent folder and check again.
+			folder = Uri.joinPath(folder, '..');
 		}
 
-		return null;
+		return false;
 	}
 
 	/**
-	 * Attempts to find an executable in the given directory and returns the path to it if found.
+	 * Attempts to find a coding standard file in the given folder and returns the path to it if found.
 	 *
-	 * @param {Uri} directory The directory we're checking for an executable in.
+	 * @param {Uri} folder The folder we're checking for a coding standard file.
 	 */
-	private async findExecutableInDirectory(
-		directory: Uri
-	): Promise<ExecutablePath | null> {
+	private async findCodingStandardFile(folder: Uri): Promise<string | false> {
+		// Look for any of the valid coding standard filenames.
+		for (const filename of AutomaticCodingStandardFilenames) {
+			try {
+				// The stat() call will throw an error if the file could not be found.
+				const codingStandardPath = Uri.joinPath(folder, filename);
+				await this.workspace.fs.stat(codingStandardPath);
+
+				return codingStandardPath.fsPath;
+			} catch (e) {
+				// Only errors from the filesystem are relevant.
+				if (!(e instanceof FileSystemError)) {
+					throw e;
+				}
+			}
+		}
+
+		return false;
+	}
+
+	/**
+	 * Attempts to find an executable in the given folder and returns the path to it if found.
+	 *
+	 * @param {Uri} folder The folder we're checking for an executable in.
+	 */
+	private async findExecutableInFolder(
+		folder: Uri
+	): Promise<ExecutablePath | false> {
 		try {
-			// We should be aware of custom vendor directories so that
+			// We should be aware of custom vendor folders so that
 			// we can find the executable in the correct location.
-			const composerPath = Uri.joinPath(directory, 'composer.json');
+			const composerPath = Uri.joinPath(folder, 'composer.json');
 
 			const composerFile = JSON.parse(
 				this.textDecoder.decode(
@@ -423,7 +555,7 @@ export class Configuration {
 
 			// Make sure to find a platform-specific executable.
 			const phpcsPath = Uri.joinPath(
-				directory,
+				folder,
 				vendorDir,
 				'bin',
 				process.platform === 'win32' ? 'phpcs.bat' : 'phpcs'
@@ -434,7 +566,7 @@ export class Configuration {
 
 			// The lack of an error indicates that the file exists.
 			return {
-				workingDirectory: directory.fsPath,
+				workingDirectory: folder.fsPath,
 				executable: phpcsPath.fsPath,
 			};
 		} catch (e) {
@@ -444,6 +576,6 @@ export class Configuration {
 			}
 		}
 
-		return null;
+		return false;
 	}
 }